requirements.txt
flask
openai
yt-dlp
pysrt
python-dotenv
